Point Plotting
Created function that will plot caffine decay after inputting a caffine
level.  It works but I had to comment out code in HomePage that depended
on the "getPoint()" function in database. I was told "getPoint()" would
be deleted anyways. IMPORTANT: It can only output the graph of one
drink, set to an initial 500 mg.