The Database
Created a database that holds 2 tables, one for consumed caffine and
another for caffine available for consumtion. The functions for
interacting with this database are in a seperate file. A 3rd table was
made to test the funcionality of the database, and it currently holds
the values of caffine decay.