Multiple Caffeine Support Part II
Mulitple Caffeinated items can now be selected, and the program accounts
for that. The only problem is the graph needs to be redrawn every time
someone drinks. Also the caffeine drink select list now has the list of
drinks in the database. Also changed AddDrink so that caffine can only
be added a certain amount of time from now